+++ src/pipecat/transports/services/daily.py
@@ -900,11 +900,11 @@ class DailyTransport(BaseTransport):
 
     async def send_image(self, frame: OutputImageRawFrame | SpriteFrame):
         if self._output:
-            await self._output.process_frame(frame, FrameDirection.DOWNSTREAM)
+            await self._output.queue_frame(frame, FrameDirection.DOWNSTREAM)
 
     async def send_audio(self, frame: OutputAudioRawFrame):
         if self._output:
-            await self._output.process_frame(frame, FrameDirection.DOWNSTREAM)
+            await self._output.queue_frame(frame, FrameDirection.DOWNSTREAM)
 
     def participants(self):
         return self._client.participants()
@@ -1033,6 +1033,8 @@ class DailyTransport(BaseTransport):
         await self._call_event_handler("on_first_participant_joined", participant)
 
     async def _on_transcription_message(self, message):
+        await self._call_event_handler("on_transcription_message", message)
+
         participant_id = ""
         if "participantId" in message:
             participant_id = message["participantId"]